requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
pandas==2.2.3
openpyxl==3.1.5
pypdf==5.3.0
python-docx==1.1.2
scikit-learn==1.5.2
python-dotenv==1.0.1
pytest==8.3.3
gunicorn==23.0.0
